The 2 down migration drops this table.

+++ framework/Service_Weather/migration/Horde/Service/Weather/1_horde_service_weather_airports.php
@@ -164,6 +164,8 @@ class HordeServiceWeatherAirports extends Horde_Db_Migration_Base
 
     public function down()
     {
-        $this->dropTable('horde_metar_airports');
+        if (in_array('horde_metar_airports', $tableList)) {
+            $this->dropTable('horde_metar_airports');
+        }
     }
 }
